Minimally working system.
Starting and stopping agents works. There is a minimal messaging layer.
I've implemented an agent lifecycle state machine which seems to work
quite well, though it can get out of sync with what is happening with
the agents.

Agents survive an agency crash and the agency can still control the
agent after it's restarted. This is not particularly robust but the
basis of a working system is in place.

Logging is still a mess and any tools (ie send) are *very* basic.