Move the state machine into a new class calld AgentProcess.
This *dramatically* reduces the amount of code; I suspect because this
is the correct way to use it!

I've added before/error callbacks for logging.

I would like to remove the state machine from the cache when an agent is
stopped but I cannot see an easy way of doing this. The way I've
implemented it doesn't work properly.

I've commented out the monitoring stuff as this doesn't really work very
well. But starting and stopping is reliable (when the state machine
doesn't get out of sync with reality).