New logger.
I've done the logger like this so the logger nows what class/line number
it's being called from. The previous way didn't allow that. It's slightly
more ackward to use but only a bit.

Having said that I'm not completely happy with the way I've implemented
it. I need to learn more about modules and variables etc. Anyway read
the comment at the top of the Logger class.