Simple smithctl. This only logs the output of the command.
I've done this using metaprogramming which means that both smithctl &
the agency need to know all the commands. This is not ideal but given
that some commands need to return data to smtithctl and others don't it
needs to know about the commands. Having said that all commands could
return a status (which is probably a good idea anyway) which would
simplify smithctl. Umm thinking aloud!