Add reap method when the state is changed to :not_responding.
I'm not sure where I'm going with this. My thinking was that if an agent
gets in an unknown state then it should be cleaned up and, if specified,
a new agent started. But this can lead to agents getting constantly
started, getting wedged, killed and then restarted. The sysadmin might
also want to have a look at the process to see what it's doing.

Maybe making it an agent option is the right thing to do, or maybe a
system wide parameter.