Change the agents can be written.
Before this commit and agent was written in the run method. This method
was called by the bootstrapper. Given that virtually all agents are
written with the run method calling get_message(queue_name) I made it
so all you need provide is a default_handler and the Agent class will
create a queue named after the agent name and attach the default_handler
to that queue.

This covers about 90% of use cases and if you need to something more
complex or need to listen on multiple queues that you can fall back to
using the run method.