Reimplement Agent#run to use Defer & add Agent.task.
I've added threading so that blocking code can be run in the agent. I've
implemented it using EM.defer and set the thred pool size to 1. I'm not
particularly happy with this; I would rather call ack in the thread so
that the subscribe method doesn't actually pull a message off the queue
but that deadlocks fairly quickly, even if I use EM.schedule. I'll send
an email to the list to see what I'm doing wrong.

However setting the internal thread pool to just contain one thread
seems to achieve the same result.

There is a small slow down of about 7% (note I did one run of 100 000
messages so take this with a pince of salt) using threads, which I'm
not too worried about that. It's still about 2100 messages per second on
my laptop and if you need that ~ 7% then you can use Agent#listen and
not use the thread pool.

Instead of using defining default_handler in the agent you now use the
Agent.task class method. This is a nicer api.