Add a block to Messaging#send_message.
This allows me to attach a callback to the publish message which is
called when the message has been sent. Though saying when a message has
been sent is a complicted issue. See:
http://rubydoc.info/github/ruby-amqp/amqp/master/file/docs/Exchanges.textile#Publishing_callback_and_reliable_delivery_in_distributed_environments
for details.

Note. You need to set the correct options for this to work properly. See
the above link for details.