Add another log level: verbose.
I've added Logging.reset to surpress a warning which is caused by
Logging.init being called more than once when the specs are run. I'm not
sure if it's due to the way I'm running the specs or if it is actually a
problem. In fact I will probably remove it.