Remove Logging.reset & replace with nasty hack!
This is slightly improved in that Logging.init isn't called multiple
times but it is a bit of a hack. I guess the correct way to call
Logging.init would be right at the begining of the program so that it is
guarenteed to called once but that means I have to worry about calling
it.