Rewrite the messaging layer.
The messaging class was very crufty and not very nice to work with. This
reimplements that. I've modeled it as a sender/receiver class which
simplifies the implementation of publish_and_receive &
subscribe_and_reply which have always been particularly messy. I'm not
sure how efficient this is but it doesn't make any more trips to the
broker than it should so it's probably not too bad.

This needs a proper audit especially with respect to the options; they
always confuse me!