All commands return their results to smithctl.
Whenever a command it sent to the agency, before this commit would
simply log the result of that command, which made commands that are
supposed to return data to smithctl painful to use. This commit fixes
that.

It is not perfect as the results are badly formatted and you cannot
change it, but until I implement full command line processing for each
command that's the best it's going to get.