Don't be so strict in Messaging::Receiver#subscribe_and_reply.
Before this would raise an exception if the reply_to wasn't set. This,
for example, means that the agency can only receive messages from
senders that set reply_to and consequently reply to. Sometimes I don't
want a reply sent in which case I'm screwed.

I'm still slightly uncertain about this but I'll leave it in.