Convert all messages to use the new Protocol Buffers encoders.
Well that was a bit of a shit fight!

To send a message you must encode the message before publishing it.
This makes the code slghtly more long winded but it ensure some safety
when passing messages; not to mention that protocol buffers is almost
certainly faster then Marshal or JSON (Need benchmark here.). Of course
the implementation probably isn't the fastest, there's almost certaily
room for improvement.