Add protocol buffer to the keepalive message.
There is problem here, in fact it's a general problem with proto
buffers and AMQP. AMQP calls to_s on all messages which works when all
the fields are strings but if any of the fields are ints then the values
are wrong. I need to send a message to the mailing list.

As a work around I've made all fields strings. Not idea but it works.