Pass a "responder" proc to the subscribe_and_reply block.
Given asynchronous nature of Smith the part of the code that returns the
data in the subscribe_and_reply block may itself be run in an async
block. Therefore it's not simply a matter of assuming the return value
is the output. By passing a proc in that deals with returning data to
the caller the proc can be called in async code in needed. It slightly
complicates the code but I think it is a much better way of doing it.