Major API refactor.
Well I say major AFI refactor but the actual API refactor is not that
big but for the code behind it this is a big change.

The channel, queue and exchange creation in Endpoint, up until this
commit, have been done without using the callback for each of these.
Thit has apprently been ok but when using the Request/Reply pattern the
destination can reply before the reply queue is created causing a 313
error.

So I've introducing (drum roll) a #ready method so when a Sender or
Receiver is instantiated you must do everything in the #ready block
otherwise there will be pain and much nashing of teeth.

It may be that this could go into the constructor but I don't actually
mind it here and I think it conveys the message about exactly what it is
doing.