Made the responder an Object. Allowing the return value to be a block.
Some of the commands are getting a bit messy so by allowing them to
return as a block makes things a log cleaner.

I am starting to wonder if I'm going a bit far with the whole responder
stuff but I'm not really sure how else to do it. I could use fibres but
that's kind of fudging it anyway (they're cool in apis but most of this
isn't an api). I'll keep an eye on it.