Add :auto_delete => false to agent Receiver classes.
Before this change all queues were auto_delete, meaning that if there is
nothing listening on the queue it would get deleted. Unfortunately this
delection would occur even if there are messages on the queue. In short
if an agent crashes and there are 1000 messages in the queue then all
the messages are removed. Not good.

This change adds set auto_delete to false meaning that if the agent dies
messages are preserved. It is also a pain in the arse! It means that if
anyone is using the Receiver class directly and they will have to set
auto_delete to false. The reason I haven't set it to false by default is
that reply messages really do need to have auto_delete set to true and
the agency doesn't need auto_delete to be set to false.

There may well be a different/better way round this but I'm going to
leave it like this for the time being as all the amqp queue/exchange
options make my head hurt and given that most people will just use the
interface provided by the Agent class then it should be ok.