Reimplement Responder class using Deferrable.
Instead of re-implementing the Deferrable class with my own just use
the one that comes with EventMachine! I've also moved Responder into
it's own file.

You might argue that I should get rid of the Responder class and include
Deferrable directly into the Receiver class but that means I have to
pass the receiver class into the callback which is big. It may be
alright to do that but I need to think about it a bit more.