Add second timeout to smithctl.
This is very hacky and needs to be properly incorporated into the
sender/receivers. So I could use Deferrable which has a timeout built in
but I need to completely change the api. If I don't use Deferrable I
still have to change the api to incorporate the errback method.

<rambling stream of concious commit message>
So it looks like I might need to change the api but that would be a pain
and it would mean that it's completely different from the amqp gem.
Though, of course, I could have a class wide error handler. Um, that might
work.</rabbli...>