Add support for raw protocol buffer files.
This commit add support for compiling protcol buffer files. There is an
option in the config to specify a path that will get searched for
protocol buffer files.

It also means that it is easy for the users of Smith to specify their
own protocol buffer files for application messages; just add the files
to a directory in specified in protocol_buffer_path.