Replace existing beefcake files with proto files.
That is all but AgenyCommaend which is required by smithctl to communicate
with the agency. I could could use proto files for everything but that
would require smithctl knowing where the proto files are which just makes
life complicated and it shouldn't really need to know where they are.

Also added agent_stats.proto.