Namespace change. Memoize the encoder/decoder classes.
This looks like a big change but is mainly due to a namespace change
form Messaging to ACL (Agent Control Language) which is much more
appropriate.

I've also added code to memoize the decoder/encoder classes and simplify
the code -- or at least made it more understandable. The upshot of this
is that Payload.new will only accept a symbol for defining the actual
ACL type. This is not bad thing as using the class was rarely used.