Move the timeout mechanism to the Endpoint class.
I still have my concerns about the way I'm doing this. Homever I've had
a play with using the timeout mechanism in Deferrable and It's pretty
messy and ultimately confusing so I will leave this as it is.

It may be a symptom of something more fundamental. Maybe it's the
secret location of the Higgs Boson. Maybe it's the right way to do it.