Better message type handling for send.
You can supply the message type and a json representation of the message
so you can send any message type. This requires that you have
protocol_buffer_cache_path set in the config. Which I'm thinking of
enforcing anyway.

Add a durable option so you can send messages to durable or non-durable
queues.