Simplify the pb cache path.
Before this patch if protocol_buffer_cache_path wasn't set it would
create a tmp directory and asign the cache path to that timpdir. This
is fine for the agency but causes problems for any thing else that needs
access to them. So if protocol_buffer_cache_path isn't set I just create
$HOME/.smith/pb and put the compiled pb files there.

This also means that agency_command dosn't need to be treated specially.