Lazily create queues. Call Endpoint#ready in the queue method.
This removes a block level, which is starting to get silly (yes I know
it's EM). This should probably go into the Receiver/Sender classes but I
really need to think about the consequees of that and, more to the
point, whether it will actually work.

I've also added a some convenience methods.

I'm not happy with the naming of this class but in lieu of anything
better I'll keep it as it is.