Updade the Agent class with the new API. Remove most wrappers.
I've exposed the queue factory into the agent, I've also written some
wrappers (receiver & sender) to make using it a bit cleaner. I removed
all the old wrappers which were quite limited.

This is slowly converging on the final agent API. It's not completely
there and there are some things that I'm not happy with but it's getting
there.

I will probably leave it where it is and start doing some real work with
it just to see how it fares.