Pass a Reply object into the Receiver#subscribe block.
Instead of passing the metadata and payload into the subsrible block I
now pass a Reply (*very* badly named - will rename when I think of a
better name) object. This object pretty just wraps the metadata &
payload objects & provides accessor methods. It does mean that I don't
have to pass the metadata object into the block which is almost never
used.

The corollary of this is a cleaner public API, on the receive side
anyway. The sending side still needs to be cleaned up but even that
isn't terrible.