Don't pass the exchange to Reply. Use #send to get internal objects.
This is a nasty abuse of Ruby's protection mechanism but it does clean
the code up a bit and given that Reply class is very tightly coupled to
the Receiver class I'm going to keep it; for the time being. I suspect
this is an indication of a more fundamental design problem. See the
FIXME in the code for more details.