Initial config agent.
A user can run the ConfigAgent (or not) and send it a configuration that
will get saved to a leveldb database. Any agent can then use this to get
agent specific config.

I have grand plans for this! In the future an agent will be able to
register for updates so that when something updates a config it is
pushed out to all instances of that agent. It is entirely up to the
agent how it deals with that new information.