Add facility to add metadata to an agent.
This is pretty basic but it does allow arbitrary metadata to be added to
an agent and then query at a later date. Anything that is a string can
go in here.

This is mainly aimed at the sysadm so they can get more information on
the agents in the system. Though obviously they are at the mercy of the
developers!