Add a print option to pop-queue and make it work properly.
Before this commit the queue was pop'ed and ack'ed/reject'ed in the main
iterator loop now the main iterator just pops each queue into an Array
and the messages are ack'ed/reject'ed in the "done" part of the
iterator.

The upshot of this is I can pop, without the remove option, multiple
messages and they will be displayed properly, instead of continuously
pop'ing the last message.