smithctl now recognise two types of command: smithctl & agency commands.
Running all commands in the agency was getting a bit messy and sometimes
it was better for agents not to be run be the agency at all. This commit
splits the commands into ones that are run in the agency and ones that
are run by smithctl directly. The actual command can run in either the
agency or smithctl.

At the moment smithcctl needs to be told whether the command is an
agency command or not. This feels wrong but given that each command
can be run in either maybe it's not. I'll probably add a method to each
command saying whether it's an agency or smithctl command.

I've also changed pop_queue to be pop, added an rm (to remove a queue)
command and moved cat and top into smithctl.