Add support for any type of appender.
This uses the config item appender and directly passes it to the
appender. This makes it versatile, in that it can use any appender
including ones that don't currently exist, but it also makes it fragile
because if you get it wrong it will explode. This last point could be
alleviated by putting some error checking in!