Refactor message replies.
Instead of having a different method (publish_and_receive) the api
exposes a on_reply method which sets everything up for the proper
replying of messages.

This improves things in two ways a) the api is much cleaner: there is
only one publish method and b) the reply queue is set up once and used
multiple times. This is a significant performance benefit, not to
mention a better use of resources in general.

This is still a work in progress but the main functionality is there. I
still need to sort out the message_id/correlation_id matching is going
to work. There is a half backed attempt in there at the moment but it
won't work.