Finally fix sender and reply API.
Ok. I think I've got it! I've reverted back to the repy callback as I
think this is probably the least confusing.

For every message sent the sender keeps track of the message_id and
matches it against the correlation_id in the reply.

This is still very raw but I'm really hoping this will be the last API
change.