Complete rewrite of the messaging subsystem.
The main change behind this patch is that queues/exchanges now their own
channel instead of sharing one. There are knock on effects of this, for
example non-ack'ed message does not block messages to other queues &
prefetch can be used on a per queue basis (wild applause).

I've also changed the implementation of reply queues which fixes the
resource leak that used to happen when replying. A reply message type
can also be specified (the reply type was always default before).

The code is much simplified and there is an API change which I think
makes the overall API more consistent.