Remove the dependency on DataMapper.
I've never been very happy about this but it was quite convenient. Now I
just manage saving the state of the state machine using state_machines
callbacks. It's probably not as elegent but it works perfectly well and
it does give me more control.

I've implement persistence using protocol buffers and leveldb (thanks
google!). I can't decide if this is quite neat or egregiously bad. I
guess time will tell!

This is a work in progress.