Use the agent UUID as the key for internal house keeping.
Instead of using the name of the agent as the lookup key

This has quite reaching implications, but it mostly means that multiple
agents of the same type can be started at the same time. It also means
that managing running agents must be done using the UUID which is
nowhere near as convenient. I have however added a -n option to stop
cammnd that will stop all agents with the supplied name. The list
command will also work with a name though it should also be changed to
have a -n option.

Changes in this commit:

* The bootstrpper now works out where to find the agent instead of
  passing the path to the bootstrapper. This reduces the string length
  of the exec call which means that there aren't so many spaces in the
  process table.
* Remove the --kill option to the start command it didn't really work.
* Split the AgentLifecycle ACL into AgentAcknowledgeStart, AgentDead &
  AgentAcknowledgeStop.
* Add an extra state to the state machine to check if there is an agent
  with the given name. I've also cleaned up the state machine in general.
* Be more aggressive about removing any dead or stopped agents from the
  agent store.