Add a state method to the agent.
This is mainly used when the agent is shutting down. For example if a
connection thorn down it may well call errbacks that need to know if the
agent is shutting down.

I set each state that the agent sees (starting, running & stopping).
Convenience methods are provided for each state.