Add general mechanism for defining queues.
This allows you to define queues using a Smith::QueueDefinition class
which means you can define the queues and options in one place and reuse
them where ever needed. Quite why it's taken me this long to implement
this is beyond me.

I should point out the the old way of defining a queue name and an
options hash still works.

I've only done this for internal queues at the moment but it won't
require much work to add it to the sender/receiver wrapper methods.