Add support for config files in the cwd.
This will check the local directory for a config file and use it if it
finds one. If not if will go back through the directory tree until it
finds one or reaches $HOME.

It would probably be useful to also check /etc, /etc/smith and specify
the config file as a command line option.