Change the way ACL types are managed.
Instead of trying to flatten the ACL type into a string and setting that
in the message header this commit changes it to use the hash (murmur32)
of the type. I've also completely remove the payload and the factory
(which is optional) has been very much simplified. For example it
doesn't decorate the ACL classes, you can now instantiate them directly
and publish them as is.

It's hard to overstate the usefulness of this, it was a pain in the arse
to have to use the Payload to create messages. It was even more painful
when you had a fully instantiated message that you had to jump through
hoops to use.

Instead of the agency being responsible for compiling ACLs each process
that uses them compiles them. This is actually required because it needs
to generate the hash of all ACLs beforehand. This might seem like a bad
idea but only the class and it's hash are kept and only classes that
have changed are actually compiled.

The generated classes are parsed (using ruby parser) and the actual ACL
name (actually it's class name) rather than the containing file is
used to pre-generate the murmur hash. This also means that when you
smithctl ACL operates on the actual ACL rather than the file containing
any number of ACLs. This process does incur a startup overhead but this
can be cached and save to hard disk.