Simplify the logger.
This has been a source of pain for a very long time. I've removed most
of the reloading code as a) it's not needed and b) it causes more grief
than it's worth.

Before this commit the logger could change the format, log level, trace
level and the appender. This is stupid so the only thing you can change
now is the log level and the trace level.

This also fixes the resource leak that opening multiple appenders
causes.