Command to dump all messages on a queue to stdout.
This is a very dangerous command hence you have to give the following
command line option: --yes-i-want-to-remove-all-messaeges-from-the-queue.
Which should protect against the obvious typos. If you are using this
from bash or zsh make sure you add a space in front of the command so it
won't be included in your history. Oh yes you'll have to configure the
shell to work this way, for zsh: setopt hist_ignore_space.