Remove the queue garbage collector.
This is was a misguided idea and caused problems with auto-recovery. So
instead of using random queue names for replies use the queue name with
a suffix of reply. Of course this will fail when there are multiple
clients, but for the timebeing this will have to do.

There is still a channel leak with this implementation because server
side channels don't get closed but it is a comparitively slow leak.