Improve error handling and reporting.
There are case where queues or exchanges don't get cleaned up properly
so when you try and remove them an error occurs and the command aborts
make it painfull if you need to remove large numbers of queues. This
adds two new options: --ignore-errors --log-errors to override this
behaviour.

I've also replaced the --verbose option with a FIXME as it needs to be
done properly.