Add a default config file.
Insead of requiring the config file to be complete even when most of the
config items are exactly the same thi commit adds an almost complete
config file which is then merge with the user's config file.

This means that it can be config with a minial config and still work
properly.