reply queue suffix defaults to a random string.
This was causing insidious errors especially when trying to crontrol the
agency. If multiple processes are sending commands to the agency (such
as riemann-smith and smithctl) then the agency would, 50% of the time,
send the reply message to the wrong queue.

This fixes the problem by insuring the reply queue name is unique.