Add support for a group directory.
Groups are implemented using a directory containing symlinks back to the
agents files. This can conflict if namespacing is being used. This
allows the group directory to be added to a directory called groups (by
default — it's configurable) so as not to conflict.

Note if the groups directory exists then this is the only place where
the groups directories will be searched.