Support multiple directories added to the $LOAD_PATH
Adds a new configuration property to support configurable directory
paths for loading shared code. This might be useful for agents which are
deployed as part of a larger codebase with reusable modules and classes
in an `app` directory.

Usage:

```toml
[agency]
lib_directories = ["lib", "app"]
```

Defaults to `["lib"]` so that the existing behaviour is preserved by
default without needing to be explicitly configured.