Fix particularly egregious process handling bug
Process.detach was being passed 0 (which basically means when I die
ensure I'm reaped — which is nonsense) because agent_process.pid isn't
set untill acknowledge_start is called which is the last step in the
agent set up.

This uses the value from fork which is how it should be done (and once
was) grumble, grumble.