Collect stats about every event that happens to an agent
This is in a fair raw but function state. It sends lifecycle events to a
single queue "cluster.events". The agents can be configured (all by
default at the moment but the intention is to turn it on by default) to
send periodic messages back to the agency which can then distribute to
any interested parties.

The longer term intention is have an agent that listens to
"cluster.events" and saves the events to a database which can then be
displayed accordingly — I've started a sister project called smith-console.
However the same mechanism can be used for other uses such as
monitoring, for example it should be fairly easy to write a Prometheus
exporter.